JumpList を使う処理を Sizer クラスに分割しました。
