+++ Size/Size.csproj
@@ -74,6 +74,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Sizer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
